requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.0

# Synthetic data
faker>=19.0.0

# Streaming simulation
kafka-python>=2.0.0

# ML / Modeling
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Dashboard
streamlit>=1.25.0

# GenAI / LLM
google-generativeai>=0.3.0
langchain>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0